+++ CMakeLists.txt
@@ -4,4 +4,4 @@ add_subdirectory(lib)
 add_subdirectory(logging)
 add_subdirectory(net)
 
-add_subdirectory(samples/common)
+add_subdirectory_ifdef(CONFIG_GOLIOTH_SAMPLES_COMMON samples/common)
