+++ samples/common/Kconfig
@@ -1,8 +1,15 @@
 #
-# Copyright (C) 2021 Golioth, Inc.
+# Copyright (C) 2021-2022 Golioth, Inc.
 #
 # SPDX-License-Identifier: Apache-2.0
 #
+
+menuconfig GOLIOTH_SAMPLES_COMMON
+	bool "Common helper code shared across multiple samples"
+	help
+	  Build and link common code that is shared across Golioth samples.
+
+if GOLIOTH_SAMPLES_COMMON
 
 config GOLIOTH_SAMPLE_SETTINGS_AUTOLOAD
 	bool "Autoamtically load settings before main()"
@@ -24,6 +31,7 @@ config GOLIOTH_SAMPLE_SETTINGS_SHELL
 config GOLIOTH_SAMPLE_WIFI
 	bool "WiFi utilities for samples"
 	depends on NET_L2_WIFI_MGMT
+	select GOLIOTH_SAMPLE_INCLUDE_HEADERS
 	help
 	  Enable utilities for easy WiFi setup, mainly for use inside samples/.
 
@@ -50,3 +58,5 @@ config GOLIOTH_SAMPLE_WIFI_PSK
 	  unsecure (open) AP.
 
 endif # (GOLIOTH_SAMPLE_WIFI && !GOLIOTH_SAMPLE_WIFI_SETTINGS) || WIFI_ESP32
+
+endif # GOLIOTH_SAMPLES_COMMON
