+++ samples/common/include/samples/common/wifi.h
