+++ samples/hello_sporadic/prj.conf
@@ -1,3 +1,5 @@
+CONFIG_GOLIOTH_SAMPLES_COMMON=y
+
 # Generic networking options
 CONFIG_NETWORKING=y
 CONFIG_NET_IPV4=y
