+++ samples/lightdb_stream/src/main.c
@@ -9,7 +9,7 @@ LOG_MODULE_REGISTER(golioth_lightdb_stream, LOG_LEVEL_DBG);
 
 #include <net/coap.h>
 #include <net/golioth/system_client.h>
-#include <net/golioth/wifi.h>
+#include <samples/common/wifi.h>
 
 #include <drivers/sensor.h>
 #include <stdlib.h>
